Create separate repository module for local store
