+++ src/qw/local_store/__init__.py
@@ -1,1 +1,1 @@
-"""Local storage for qw."""
+"""Interacting with qw local storage, local repository and keychain."""
